Renamed wfs to fsLoader
