+++ README.md
@@ -1,14 +1,13 @@
 # wunderDB v2
 
-Hi! Welcome to the wunderDB Repository. 
+Hi! Welcome to the wunderDB Repository.
 
-Currently I am rewriting the whole project in Golang with standard API following standard development practices. So this branch is still a Work in Progress! 👷🏻 🏗️ Hope to deliver the initial version of v2 soon! 
+Currently I am rewriting the whole project in Golang with standard API following standard development practices. So this branch is still a Work in Progress! 👷🏻 🏗️ Hope to deliver the initial version of v2 soon!
 
 ## WDB-Retro
 
-The first version, based on Python Flask will not be phased out any time soon. To keep it accessible and so that the version 1 doesn't get lost in the version list, I have moved the v1 to a new repository here - [wdb-retro](https://github.com/TanmoySG/wdb-retro). 
+The first version, based on Python Flask will not be phased out any time soon. To keep it accessible and so that the version 1 doesn't get lost in the version list, I have moved the v1 to a new repository here - [wdb-retro](https://github.com/TanmoySG/wdb-retro).
 
-The wdb-retro repo has all the version 1 code as well as the Docker Image with the new name - so that when I publish the wunderDB v2 Docker Images are published the v1 Image doesnt get lost and anyone planning to use that still can. The version 1 will not be actively maintained. 
+The wdb-retro repo has all the version 1 code as well as the Docker Image with the new name - so that when I publish the wunderDB v2 Docker Images are published the v1 Image doesnt get lost and anyone planning to use that still can. The version 1 will not be actively maintained.
 
 Check out the final release of v1 <https://github.com/TanmoySG/wdb-retro/releases/tag/v1.1.0>
-
